+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Seminarassignment2025">
         <activity
-            android:name=".ProfileActivity"
+            android:name=".views.profile.ProfileActivity"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
