+++ app/src/main/java/com/example/seminar_assignment_2025/data/movie/MovieRepository.kt
@@ -1,4 +1,6 @@
-package com.example.seminar_assignment_2025.data
+package com.example.seminar_assignment_2025.data.movie
+
+import com.example.seminar_assignment_2025.domainmodel.Movie
 
 interface MovieRepository {
     suspend fun searchByTitle(query: String): List<Movie>
