--- app/src/main/java/com/example/seminar_assignment_2025/di/DataStoreModule.kt
@@ -1,21 +1,0 @@
-package com.example.seminar_assignment_2025.di
-
-import android.content.Context
-import com.example.seminar_assignment_2025.data.SearchHistoryRepository
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.qualifiers.ApplicationContext
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object DataStoreModule {
-
-    @Provides
-    @Singleton
-    fun provideSearchHistoryRepository(@ApplicationContext context: Context): SearchHistoryRepository {
-        return SearchHistoryRepository(context)
-    }
-}
