+++ app/src/main/java/com/example/seminar_assignment_2025/di/RepositoryModule.kt
@@ -1,7 +1,9 @@
 package com.example.seminar_assignment_2025.di
 
-import com.example.seminar_assignment_2025.data.MovieRepository
-import com.example.seminar_assignment_2025.data.MovieRepositoryImpl
+import com.example.seminar_assignment_2025.data.movie.MovieRepository
+import com.example.seminar_assignment_2025.data.movie.MovieRepositoryImpl
+import com.example.seminar_assignment_2025.data.search.SearchHistoryRepository
+import com.example.seminar_assignment_2025.data.search.SearchHistoryRepositoryImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -15,4 +17,8 @@ abstract class RepositoryModule {
     @Binds
     @Singleton
     abstract fun bindMovieRepository(movieRepositoryImpl: MovieRepositoryImpl): MovieRepository
+
+    @Binds
+    @Singleton
+    abstract fun bindSearchHistoryRepository(searchHistoryRepositoryImpl: SearchHistoryRepositoryImpl): SearchHistoryRepository
 }
