+++ app/src/main/java/com/example/seminar_assignment_2025/domainmodel/Movie.kt
@@ -1,4 +1,4 @@
-package com.example.seminar_assignment_2025.data
+package com.example.seminar_assignment_2025.domainmodel
 
 import kotlinx.serialization.Serializable
