+++ app/src/main/java/com/example/seminar_assignment_2025/views/detail/MovieDetailViewModel.kt
@@ -1,14 +1,16 @@
-package com.example.seminar_assignment_2025.ui.detail
+package com.example.seminar_assignment_2025.views.detail
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.seminar_assignment_2025.data.Movie
-import com.example.seminar_assignment_2025.data.MovieRepository
+import com.example.seminar_assignment_2025.domainmodel.Movie
+import com.example.seminar_assignment_2025.data.movie.MovieRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
 @HiltViewModel
@@ -21,7 +23,10 @@ class MovieDetailViewModel @Inject constructor(
 
     fun getMovieDetail(movieId: Int) {
         viewModelScope.launch {
-            _movieDetail.value = movieRepository.getMovieDetail(movieId)
+            val detail = withContext(Dispatchers.IO) {
+                movieRepository.getMovieDetail(movieId)
+            }
+            _movieDetail.value = detail
         }
     }
