+++ app/src/main/java/com/example/seminar_assignment_2025/views/game/GameScreen.kt
@@ -1,4 +1,4 @@
-package com.example.seminar_assignment_2025.ui
+package com.example.seminar_assignment_2025.views.game
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.gestures.detectDragGestures
@@ -19,10 +19,8 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.input.pointer.pointerInput
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.seminar_assignment_2025.game.GameViewmodel
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.setValue
