+++ app/src/main/java/com/example/seminar_assignment_2025/views/profile/ProfileActivity.kt
@@ -1,17 +1,13 @@
-package com.example.seminar_assignment_2025
+package com.example.seminar_assignment_2025.views.profile
 
-import android.content.Intent
-import android.net.Uri
 import android.os.Bundle
 import android.widget.ImageView
 import android.widget.TextView
-import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
-import androidx.core.net.toUri
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.example.seminar_assignment_2025.GithubIdAdapter
+import com.example.seminar_assignment_2025.R
 
 class ProfileActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
